main.py에서 표만드는 파일 실행
